[All] Rebuilt GUI in Python, Tweaked Auto-Encoder, Changed collectors.py in relation to SSL
GUI Rebuild Justification:
The main concern with building the GUI in Java is the lack of control and usability of the new Tensorflow library at the time of this software's development.
Since the model needs to be retrained on each new set of testing data, Java's Tensorflow library is not able to solve this problem, as it lacks the ability to re-train models made in Python.
As such, to make this platform deployable to teachers who may or may not have a Python version with the proper libraries already installed, the module PyInstaller will be used to build executables from
the python scripts. This was originally avoided, given that Python is already a comparably slower language and packaging it into a .exe file slows it down even more.

Major Changes:
[predictors.py:61:143-149]
The AutoEncoder.seperate() method now modifies and updates the GUI if a controller(tkinter.Tk() object) is passed to it, as well as a variable to modify.

[displayers.py:480]
Created a EULA agreement implementation in displayers.py in accordance with MIT license

[displayers.py:171:175]
Allowed users to turn off SSL certificate verification. This is done for teachers who may have invalid middle-man network security protection, such as a untrusted Zscaler certificate.
See README.md for more information.
Further testing is required to find a alternate to troublesome CAs without disabling ssl verification entirely, for now this is a placeholder solution.

- Updated source file to include testing from private data-sets. Will be changed in the next commit.

Minor Changes:
[LICENSE:3]
Updated years to better reflect project.

[README]
Updated to remove spelling error and flesh out information

- Updated Defualt values for predictors.AutoEncoder().seperate() to better reflect preformance tests
- Update all files that output information onto the hard drive to automatically create directories so that users won't have to be bothered with it
- Added minor function into the python collector file taken from the previous Java collector file that gets all the modules associated with a course.
- Added a UserCollector class taken from the Java version to query all courses associated with the current api user.
- Commented out building the dataframe with scores, as through testing it appears cheaters are signifigantly less likly to cheat on the entire test rather than just a few questions
- Removed any users from the dataset with a page_leaves value higher than 10 [to be updated to a better system involving taking the number of questions of a test into account]
- Updated visualizers to add a class that works with graphing output from the AutoEncoder class.
- Moved old Java GUI files into /GUI/legacy folder, to be removed in the next commit.
- Updated GitIgnore to reflect new system for making files