[GUI | KMeans] Added TTK Styling
KMeans:
Updated to include distance metrics to both the nearest and farthest away cluster point (in this case class 0 and 1). This should help distinguish false
flags, positive or negative. The intuition here, I believe, would be someone who was flagged as a cheater would have a distance to the cheating cluster
that would be greater than most distances to cluster and have a distance to the non-cheating cluster that would be less than other users flagged as cheaters.
Conversely, a cheater flagged as a non-cheater would have a larger distance to the non-cheater center and a smaller distance to the cheater center compared
to other non-cheaters.

GUI:
Separated normal displayers and displayers relating to machine learning for cleaner code
Removed legacy Java files
Added documentation virtually everywhere
Added a logo - More work to be done here.
Created a functionless settings class to be retooled as a child of a generalized settings class
Created a menu to interact with AutoEncoder settings before activating it. (Ctrl + F12 before starting the separation process).