CSRF Protection
- Enabled some more secure features of Django to ensure security
from cross site request forgery. This isn't really adding all that much
to this website as this is not the main security threat. It's just nice
to have in general. There are only GET requests to canvas, so a CSRF attack
would do nothing in this context.

- Updated logo - it's still being worked on. I think it looks much better
as the ico though.

- Added a new template that will be the redirect when the model is being
trained. However, drawing with a canvas and then animating that is quite
difficult. Looking into js libraries that take care of this.

- Put a linear layer on the autoencoder at the end just like before.